% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterYts.R
\name{filterYts}
\alias{filterYts}
\title{Filter YTS Data}
\usage{
filterYts(data, location = NULL, TopicDesc = NULL, MeasureDesc = NULL)
}
\arguments{
\item{data}{A data frame containing YTS data.}

\item{location}{A character vector specifying the location abbreviation(s) to filter by.}

\item{topic}{A character vector specifying the topics(s) to filter by.}

\item{measure}{A character vector specifying the measures(s) to filter by.}
}
\value{
A filtered data frame containing YTS data that meets the specified criteria.
data("yts_df)
}
\description{
Filter the Youth Tobacco Survey (YTS) dataset based on specified criteria.
}
