% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYts.R
\name{plotYts}
\alias{plotYts}
\title{Plot YTS Data}
\usage{
plotYts(data, x_col, y_col)
}
\arguments{
\item{data}{A data frame containing the YTS data.}

\item{x_col}{The column to plot on the x-axis.}

\item{y_col}{The column to plot on the y-axis.}

\item{title}{The title of the plot.}
}
\value{
A simple plot based on the user input.
}
\description{
Generate a simple plot based on user input.
}
\examples{
# Load the YTS data
data(yts_df)

# Plot the distribution of responses for "Cigarette Use (Youth)"
plot_yts_data(yts_df, "Year", "Data_Value", "Tobacco Consumption Over the Years")

}
