% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytsr-package.R
\docType{package}
\name{ytsr-package}
\alias{ytsr}
\alias{ytsr-package}
\title{ytsr: Youth Tobacco Survey (YTS) Data Analysis Package}
\description{
The Youth Tobacco Survey (YTS) Data Analysis Package provides tools and functions for analyzing and visualizing survey data related to tobacco use among youth. It includes functions for data cleaning, manipulation, and visualization, allowing users to explore trends, demographics, and responses from the YTS dataset.
}
\author{
\strong{Maintainer}: Eric Zarycki \email{ericdzarycki@gmail.com} (\href{https://orcid.org/0009-0001-3734-2675}{ORCID})

}
\keyword{internal}
